Moved views to admin.
