+++ admin/routes/views/home.js
@@ -1,7 +1,7 @@
-var keystone = require('../../');
+var keystone = require('../../../');
 
 exports = module.exports = function(req, res) {
-	
+
 	keystone.render(req, res, 'home', {
 		section: 'home',
 		page: 'home',
