+++ admin/routes/views/list.js
@@ -1,41 +1,41 @@
-var keystone = require('../../'),
+var keystone = require('../../../'),
 	_ = require('underscore'),
 	querystring = require('querystring');
 
 exports = module.exports = function(req, res) {
-	
+
 	var viewLocals = {
 		validationErrors: {},
 		showCreateForm: _.has(req.query, 'new')
 	};
-	
+
 	var sort = { by: req.query.sort || req.list.defaultSort },
 		filters = req.list.processFilters(req.query.q),
 		cleanFilters = {},
 		queryFilters = req.list.getSearchFilters(req.query.search, filters),
 		columns = (req.query.cols) ? req.list.expandColumns(req.query.cols) : req.list.defaultColumns;
-	
+
 	_.each(filters, function(filter, path) {
 		cleanFilters[path] = _.omit(filter, 'field');
 	});
-	
+
 	if (sort.by) {
-		
+
 		sort.inv = sort.by.charAt(0) === '-';
 		sort.path = (sort.inv) ? sort.by.substr(1) : sort.by;
 		sort.field = req.list.fields[sort.path];
-		
+
 		var clearSort = function() {
 			delete req.query.sort;
 			var qs = querystring.stringify(req.query);
 			return res.redirect(req.path + ((qs) ? '?' + qs : ''));
 		};
-		
+
 		// clear the sort query value if it is the default sort value for the list
 		if (req.query.sort === req.list.defaultSort) {
 			return clearSort();
 		}
-		
+
 		if (sort.field) {
 			// the sort is set to a field, use its label
 			sort.label = sort.field.label;
@@ -50,15 +50,15 @@ exports = module.exports = function(req, res) {
 			// it looks like an invalid path has been specified (no matching field), so clear the sort
 			return clearSort();
 		}
-		
+
 	}
-	
+
 	var renderView = function() {
-		
+
 		var query = req.list.paginate({ filters: queryFilters, page: req.params.page, perPage: req.list.get('perPage') }).sort(sort.by);
-		
+
 		req.list.selectColumns(query, columns);
-		
+
 		var link_to = function(params) {
 			var p = params.page || '';
 			delete params.page;
@@ -72,27 +72,27 @@ exports = module.exports = function(req, res) {
 			params = querystring.stringify(_.defaults(params, queryParams));
 			return '/keystone/' + req.list.path + (p ? '/' + p : '') + (params ? '?' + params : '');
 		};
-		
+
 		query.exec(function(err, items) {
-			
+
 			if (err) {
 				console.log(err);
 				return res.status(500).send('Error querying items:<br><br>' + JSON.stringify(err));
 			}
-			
+
 			// if there were results but not on this page, reset the page
 			if (req.params.page && items.total && !items.results.length) {
 				return res.redirect('/keystone/' + req.list.path);
 			}
-			
+
 			// go straight to the result if there was a search, and only one result
 			if (req.query.search && items.total === 1 && items.results.length === 1) {
 				return res.redirect('/keystone/' + req.list.path + '/' + items.results[0].id);
 			}
-			
+
 			var download_link = '/keystone/download/' + req.list.path,
 				downloadParams = {};
-				
+
 			if (req.query.q) {
 				downloadParams.q = req.query.q;
 			}
@@ -102,15 +102,15 @@ exports = module.exports = function(req, res) {
 			if (req.query.cols) {
 				downloadParams.cols = req.query.cols;
 			}
-			
+
 			downloadParams = querystring.stringify(downloadParams);
-			
+
 			if (downloadParams) {
 				download_link += '?' + downloadParams;
 			}
-			
+
 			var appName = keystone.get('name') || 'Keystone';
-			
+
 			keystone.render(req, res, 'list', _.extend(viewLocals, {
 				section: keystone.nav.by.list[req.list.key] || {},
 				title: appName + ': ' + req.list.plural,
@@ -127,11 +127,11 @@ exports = module.exports = function(req, res) {
 				submitted: req.body || {},
 				query: req.query
 			}));
-			
+
 		});
-	
+
 	};
-	
+
 	var checkCSRF = function() {
 		var pass = keystone.security.csrf.validate(req);
 		if (!pass) {
@@ -140,12 +140,12 @@ exports = module.exports = function(req, res) {
 		}
 		return pass;
 	};
-	
+
 	var item;
 	if ('update' in req.query) {
-		
+
 		if (!checkCSRF()) return renderView();
-		
+
 		(function() {
 			var data = null;
 			if (req.query.update) {
@@ -167,19 +167,19 @@ exports = module.exports = function(req, res) {
 				res.redirect('/keystone/' + req.list.path);
 			});
 		})();
-		
+
 	} else if (!req.list.get('nodelete') && req.query['delete']) { //eslint-disable-line dot-notation
-		
+
 		if (!checkCSRF()) return renderView();
-		
+
 		if (req.query['delete'] === req.user.id) { //eslint-disable-line dot-notation
 			req.flash('error', 'You can\'t delete your own ' + req.list.singular + '.');
 			return renderView();
 		}
-		
+
 		req.list.model.findById(req.query['delete']).exec(function (err, item) { //eslint-disable-line dot-notation
 			if (err || !item) return res.redirect('/keystone/' + req.list.path);
-			
+
 			item.remove(function (err) {
 				if (err) {
 					console.log('Error deleting ' + req.list.singular);
@@ -191,16 +191,16 @@ exports = module.exports = function(req, res) {
 				res.redirect('/keystone/' + req.list.path);
 			});
 		});
-		
+
 		return;
-		
+
 	} else if (!req.list.get('nocreate') && req.list.get('autocreate') && _.has(req.query, 'new')) {
-		
+
 		if (!checkCSRF()) return renderView();
-		
+
 		item = new req.list.model();
 		item.save(function(err) {
-			
+
 			if (err) {
 				console.log('There was an error creating the new ' + req.list.singular + ':');
 				console.log(err);
@@ -210,25 +210,25 @@ exports = module.exports = function(req, res) {
 				req.flash('success', 'New ' + req.list.singular + ' ' + req.list.getDocumentName(item) + ' created.');
 				return res.redirect('/keystone/' + req.list.path + '/' + item.id);
 			}
-			
+
 		});
-		
+
 	} else if (!req.list.get('nocreate') && req.method === 'POST' && req.body.action === 'create') {
-		
+
 		if (!checkCSRF()) return renderView();
-		
+
 		item = new req.list.model();
 		var updateHandler = item.getUpdateHandler(req);
-		
+
 		viewLocals.showCreateForm = true; // always show the create form after a create. success will redirect.
-		
+
 		if (req.list.nameIsInitial) {
 			if (!req.list.nameField.validateInput(req.body, true, item)) {
 				updateHandler.addValidationError(req.list.nameField.path, req.list.nameField.label + ' is required.');
 			}
 			req.list.nameField.updateItem(item, req.body);
 		}
-		
+
 		updateHandler.process(req.body, {
 			// flashErrors: true,
 			logErrors: true,
@@ -241,10 +241,10 @@ exports = module.exports = function(req, res) {
 			req.flash('success', 'New ' + req.list.singular + ' ' + req.list.getDocumentName(item) + ' created.');
 			return res.redirect('/keystone/' + req.list.path + '/' + item.id);
 		});
-		
+
 	} else {
-		
+
 		renderView();
-		
+
 	}
 };
