+++ lib/core/routes.js
@@ -42,8 +42,8 @@ function routes(app) {
 			app.all('/keystone*', this.session.persist);
 		}
 
-		app.all('/keystone/signin', require('../../routes/views/signin'));
-		app.all('/keystone/signout', require('../../routes/views/signout'));
+		app.all('/keystone/signin', require('../../admin/routes/views/signin'));
+		app.all('/keystone/signout', require('../../admin/routes/views/signout'));
 		app.all('/keystone*', this.session.keystoneAuth);
 
 	} else if ('function' === typeof this.get('auth')) {
@@ -67,7 +67,7 @@ function routes(app) {
 	}
 
 	debug('setting keystone Admin Route');
-	app.all('/keystone', require('../../routes/views/home'));
+	app.all('/keystone', require('../../admin/routes/views/home'));
 
 	// Email test routes
 	if (this.get('email tests')) {
@@ -93,8 +93,8 @@ function routes(app) {
 
 	// Admin-UI API
 	debug('setting list and item details admin routes');
-	app.all('/keystone/:list/:page([0-9]{1,5})?', initList(true), require('../../routes/views/list'));
-	app.all('/keystone/:list/:item', initList(true), require('../../routes/views/item'));
+	app.all('/keystone/:list/:page([0-9]{1,5})?', initList(true), require('../../admin/routes/views/list'));
+	app.all('/keystone/:list/:item', initList(true), require('../../admin/routes/views/item'));
 
 	return this;
 }
